refactor: extract catalog items route
